requirements.txt
Babel==2.6.0
bleach==3.1.1
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.10
django-activity-stream==0.7.0
django-bootstrap==0.2.4
django-bower==5.2.0
django-classy-tags==0.8.0
django-cookie-law==2.0.1
django-datetime-widget==0.9.3
django-debug-toolbar==1.10.1
django-djconfig==0.8.0
django-extensions==2.1.4
django-filter==2.0.0
django-fontawesome==1.0
django-fontawesome-5==1.0.16
django-haystack==2.8.1
django-hitcount==1.3.1
django-infinite-scroll-pagination==0.2.0
django-js-asset==1.2.2
django-jsonfield==1.2.0
django-jsonfield-compat==0.4.4
django-leaflet==0.24.0
django-model-utils==3.1.2
django-mptt==0.9.1
django-nyt==1.1.3
django-phonenumber-field==2.2.0
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-scheduler==0.8.8
django-sekizai==1.0.0
django-simple-captcha==0.5.10
django-stdimage==4.0.0
django-summernote==0.8.11.4
django-taggit==1.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
gunicorn==19.9.0
html5lib==1.0.1
icalendar==4.0.3
idna==2.7
Markdown==2.6.11
mistune==0.8.4
olefile==0.45.1
phonenumbers==8.10.4
Pillow==6.2.1
progressbar2==3.38.0
psycopg2-binary==2.7.7
pytest-runner==4.2
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
PyYAML==5.1
requests==2.20.1
six==1.11.0
sorl-thumbnail==12.5.0
sqlparse==0.2.4
uni-slugify==0.1.4
Unidecode==1.0.22
urllib3==1.24.2
webencodings==0.5.1
whitenoise==4.1.2
Whoosh==2.7.4
